# perf(blooms): Avoid tiny string allocations for insert cache